Split chapel.cpp into many files to make it more manageable
