+++ tools/chapel-py/src/chapel.cpp
@@ -21,692 +21,9 @@
 #include "Python.h"
 #include "chpl/framework/Context.h"
 #include "chpl/parsing/parsing-queries.h"
+#include "iterator-support.h"
+#include "core-types.h"
 #include <utility>
-
-struct IterAdapterBase {
-  virtual ~IterAdapterBase() = default;
-  virtual const chpl::uast::AstNode* next() = 0;
-};
-
-template <typename IterPair>
-struct IterAdapter : IterAdapterBase {
- private:
-  using IterType = decltype(std::declval<IterPair>().begin());
-  IterType current;
-  IterType end;
-
- public:
-  IterAdapter(IterPair pair) : current(pair.begin()), end(pair.end()) {}
-
-  const chpl::uast::AstNode* next() override {
-    if (current == end) return nullptr;
-    return *(current++);
-  }
-};
-
-static PyTypeObject* parentTypeFor(chpl::uast::asttags::AstTag tag);
-
-typedef struct {
-  PyObject_HEAD
-  IterAdapterBase* iterAdapter;
-  PyObject* contextObject;
-} AstIterObject;
-extern PyTypeObject AstIterType;
-
-typedef struct {
-  PyObject_HEAD
-  int current;
-  int num;
-  const void* container;
-  chpl::UniqueString (*nameGetter)(const void*, int);
-  const chpl::uast::AstNode* (*childGetter)(const void*, int);
-  PyObject* contextObject;
-} AstCallIterObject;
-extern PyTypeObject AstCallIterType;
-
-typedef struct {
-  PyObject_HEAD
-  chpl::Context context;
-  /* Type-specific fields go here. */
-} ContextObject;
-extern PyTypeObject ContextType;
-
-typedef struct {
-  PyObject_HEAD
-  chpl::Location location;
-} LocationObject;
-extern PyTypeObject LocationType;
-
-typedef struct {
-  PyObject_HEAD
-  PyObject* contextObject;
-  const chpl::uast::AstNode* astNode;
-} AstNodeObject;
-extern PyTypeObject AstNodeType;
-
-/**
-  Declare a Python PyTypeObject that corresponds to an AST node with the given
-  name and tag. The tag is not the same as the name because abstract base
-  classes like NamedDecl have corresponding tags called START_NamedDecl
-  and END_NamedDecl, but not NamedDecl.
- */
-#define DEFINE_PY_OBJECT_FOR(NAME, TAG)\
-  typedef struct { \
-    AstNodeObject parent; \
-  } NAME##Object; \
-  \
-  extern PyTypeObject NAME##Type; \
-  static int NAME##Object_init(NAME##Object* self, PyObject* args, PyObject* kwargs) { \
-    return parentTypeFor(chpl::uast::asttags::TAG)->tp_init((PyObject*) self, args, kwargs); \
-  } \
-
-/* Generate a Python object for reach AST node type. */
-#define AST_NODE(NAME) DEFINE_PY_OBJECT_FOR(NAME, NAME)
-#define AST_LEAF(NAME) DEFINE_PY_OBJECT_FOR(NAME, NAME)
-#define AST_BEGIN_SUBCLASSES(NAME) DEFINE_PY_OBJECT_FOR(NAME, START_##NAME)
-#define AST_END_SUBCLASSES(NAME)
-#include "chpl/uast/uast-classes-list.h"
-#undef AST_NODE
-#undef AST_LEAF
-#undef AST_BEGIN_SUBCLASSES
-#undef AST_END_SUBCLASSES
-
-/**
-  Create a Python object of the class corresponding to the given AST node's
-  type. For example, an Identifier node will be wrapped in a a chapel.Identifier.
- */
-static PyObject* wrapAstNode(ContextObject* context, const chpl::uast::AstNode* node) {
-  PyObject* toReturn = nullptr;
-  if (node == nullptr) {
-    Py_RETURN_NONE;
-  }
-  PyObject* args = Py_BuildValue("(O)", (PyObject*) context);
-  switch (node->tag()) {
-#define CAST_TO(NAME) \
-    case chpl::uast::asttags::NAME: \
-      toReturn = PyObject_CallObject((PyObject*) &NAME##Type, args); \
-      ((NAME##Object*) toReturn)->parent.astNode = node->to##NAME(); \
-      break;
-#define AST_NODE(NAME) CAST_TO(NAME)
-#define AST_LEAF(NAME) CAST_TO(NAME)
-#define AST_BEGIN_SUBCLASSES(NAME) /* No need to handle abstract parent classes. */
-#define AST_END_SUBCLASSES(NAME)
-#include "chpl/uast/uast-classes-list.h"
-#undef AST_NODE
-#undef AST_LEAF
-#undef AST_BEGIN_SUBCLASSES
-#undef AST_END_SUBCLASSES
-    default: break;
-  }
-  Py_XDECREF(args);
-  return toReturn;
-}
-
-static PyObject* wrapIterAdapter(ContextObject* context, IterAdapterBase* iterAdapter) {
-  auto argList = Py_BuildValue("(O)", (PyObject*) context);
-  auto astIterObjectPy = PyObject_CallObject((PyObject *) &AstIterType, argList);
-  auto astIterObject = (AstIterObject*) astIterObjectPy;
-
-  astIterObject->iterAdapter = iterAdapter;
-
-  Py_XDECREF(argList);
-  return astIterObjectPy;
-}
-
-template <typename IterPair>
-static IterAdapterBase* mkIterPair(const IterPair& pair) {
-  return new IterAdapter<decltype(pair)>(pair);
-}
-
-template <typename IterPair>
-static PyObject* wrapIterPair(ContextObject* context, const IterPair& pair) {
-  return wrapIterAdapter(context, new IterAdapter<decltype(pair)>(pair));
-}
-
-static PyTypeObject* parentTypeFor(chpl::uast::asttags::AstTag tag) {
-#define AST_NODE(NAME)
-#define AST_LEAF(NAME)
-#define AST_BEGIN_SUBCLASSES(NAME)
-#define AST_END_SUBCLASSES(NAME) \
-  if (tag > chpl::uast::asttags::START_##NAME && tag < chpl::uast::asttags::END_##NAME) { \
-    return &NAME##Type; \
-  }
-#include "chpl/uast/uast-classes-list.h"
-#include "chpl/uast/uast-classes-list.h"
-#undef AST_NODE
-#undef AST_LEAF
-#undef AST_BEGIN_SUBCLASSES
-#undef AST_END_SUBCLASSES
-  return &AstNodeType;
-}
-
-static int AstIterObject_init(AstIterObject* self, PyObject* args, PyObject* kwargs) {
-  PyObject* astObjectPy;
-  if (!PyArg_ParseTuple(args, "O", &astObjectPy))
-      return -1;
-
-  auto contextObject = (ContextObject*) astObjectPy;
-
-  Py_INCREF(contextObject);
-  self->contextObject = (PyObject*) contextObject;
-  self->iterAdapter = nullptr;
-
-  return 0;
-}
-
-static void AstIterObject_dealloc(AstIterObject* self) {
-  delete self->iterAdapter;
-  Py_XDECREF(self->contextObject);
-  Py_TYPE(self)->tp_free((PyObject *) self);
-}
-
-static PyObject* AstIterObject_iter(AstIterObject *self) {
-  Py_INCREF(self);
-  return (PyObject*) self;
-}
-
-static PyObject* AstIterObject_next(AstIterObject *self) {
-  if (self->iterAdapter) {
-    if (auto nextNode = self->iterAdapter->next()) {
-      return wrapAstNode((ContextObject*) self->contextObject, nextNode);
-    }
-  }
-  PyErr_SetNone(PyExc_StopIteration);
-  return nullptr;
-}
-
-PyTypeObject AstIterType = {
-  PyVarObject_HEAD_INIT(NULL, 0)
-  .tp_name = "AstIter",
-  .tp_basicsize = sizeof(AstIterObject),
-  .tp_itemsize = 0,
-  .tp_dealloc = (destructor) AstIterObject_dealloc,
-  .tp_flags = Py_TPFLAGS_DEFAULT,
-  .tp_doc = PyDoc_STR("An iterator over Chapel AST nodes"),
-  .tp_iter = (getiterfunc) AstIterObject_iter,
-  .tp_iternext = (iternextfunc) AstIterObject_next,
-  .tp_init = (initproc) AstIterObject_init,
-  .tp_new = PyType_GenericNew,
-};
-
-static int AstCallIterObject_init(AstCallIterObject* self, PyObject* args, PyObject* kwargs) {
-  PyObject* astObjectPy;
-  if (!PyArg_ParseTuple(args, "O", &astObjectPy))
-      return -1;
-  auto astObject = (AstNodeObject*) astObjectPy;
-
-  Py_INCREF(astObject->contextObject);
-  self->contextObject = astObject->contextObject;
-
-  return 0;
-}
-
-static void AstCallIterObject_dealloc(AstCallIterObject* self) {
-  Py_XDECREF(self->contextObject);
-  Py_TYPE(self)->tp_free((PyObject *) self);
-}
-
-static PyObject* AstCallIterObject_iter(AstCallIterObject *self) {
-  Py_INCREF(self);
-  return (PyObject*) self;
-}
-
-static PyObject* AstCallIterObject_next(AstCallIterObject *self) {
-  if (self->current == self->num) {
-    PyErr_SetNone(PyExc_StopIteration);
-    return nullptr;
-  }
-  auto argName = self->nameGetter(self->container, self->current);
-  auto child = wrapAstNode((ContextObject*) self->contextObject,
-                           self->childGetter(self->container, self->current));
-  PyObject* toReturn = nullptr;
-  if (!argName.isEmpty()) {
-    toReturn = Py_BuildValue("sO", argName.c_str(), child);
-    Py_XDECREF(child);
-  } else {
-    toReturn = child;
-  }
-  self->current++;
-  return toReturn;
-}
-
-PyTypeObject AstCallIterType = {
-  PyVarObject_HEAD_INIT(NULL, 0)
-  .tp_name = "AstCallIter",
-  .tp_basicsize = sizeof(AstCallIterObject),
-  .tp_itemsize = 0,
-  .tp_dealloc = (destructor) AstCallIterObject_dealloc,
-  .tp_flags = Py_TPFLAGS_DEFAULT,
-  .tp_doc = PyDoc_STR("An iterator over Chapel function call actuals"),
-  .tp_iter = (getiterfunc) AstCallIterObject_iter,
-  .tp_iternext = (iternextfunc) AstCallIterObject_next,
-  .tp_init = (initproc) AstCallIterObject_init,
-  .tp_new = PyType_GenericNew,
-};
-
-static int LocationObject_init(LocationObject* self, PyObject* args, PyObject* kwargs) {
-  new (&self->location) chpl::Location();
-  return 0;
-}
-
-static void LocationObject_dealloc(LocationObject* self) {
-  self->location.~Location();
-  Py_TYPE(self)->tp_free((PyObject *) self);
-}
-
-static PyObject* LocationObject_start(LocationObject *self, PyObject* Py_UNUSED(args)) {
-  auto& location = self->location;
-  return Py_BuildValue("ii", location.firstLine(), location.firstColumn());
-}
-
-static PyObject* LocationObject_end(LocationObject *self, PyObject* Py_UNUSED(args)) {
-  auto& location = self->location;
-  return Py_BuildValue("ii", location.lastLine(), location.lastColumn());
-}
-
-static PyObject* LocationObject_path(LocationObject *self, PyObject* Py_UNUSED(args)) {
-  return Py_BuildValue("s", self->location.path().c_str());
-}
-
-static PyMethodDef LocationObject_methods[] = {
-  { "start", (PyCFunction) LocationObject_start, METH_VARARGS, "Get the start of a Location object" },
-  { "end", (PyCFunction) LocationObject_end, METH_VARARGS, "Get the end of a Location object" },
-  { "path", (PyCFunction) LocationObject_path, METH_VARARGS, "Get the path of a Location object" },
-  {NULL, NULL, 0, NULL}  /* Sentinel */
-};
-
-PyTypeObject LocationType = {
-  PyVarObject_HEAD_INIT(NULL, 0)
-  .tp_name = "Location",
-  .tp_basicsize = sizeof(LocationObject),
-  .tp_itemsize = 0,
-  .tp_dealloc = (destructor) LocationObject_dealloc,
-  .tp_flags = Py_TPFLAGS_DEFAULT,
-  .tp_doc = PyDoc_STR("The Chapel context object that tracks various frontend state"),
-  .tp_methods = LocationObject_methods,
-  .tp_init = (initproc) LocationObject_init,
-  .tp_new = PyType_GenericNew,
-};
-
-template <typename CppType>
-struct PythonReturnTypeInfo {};
-
-#define DEFINE_RETURN_TYPE(TYPE, TYPESTR, WRAP, UNWRAP) \
-  template <> \
-  struct PythonReturnTypeInfo<TYPE> { \
-    static constexpr const char* TypeString = TYPESTR; \
-  \
-    static PyObject* wrap(ContextObject* CONTEXT, const std::remove_const<TYPE>::type& TO_WRAP) { \
-      return WRAP; \
-    } \
-  \
-    TYPE wrap(ContextObject* CONTEXT, PyObject* TO_UNWRAP) { \
-      return UNWRAP; \
-    } \
-  \
-  }
-
-DEFINE_RETURN_TYPE(bool, "bool", Py_BuildValue("b", TO_WRAP), PyLong_AsLong(TO_UNWRAP));
-DEFINE_RETURN_TYPE(int, "int", Py_BuildValue("i", TO_WRAP), PyLong_AsLong(TO_UNWRAP));
-DEFINE_RETURN_TYPE(const char*, "str", Py_BuildValue("s", TO_WRAP), PyUnicode_AsUTF8(TO_UNWRAP));
-DEFINE_RETURN_TYPE(chpl::UniqueString, "str", Py_BuildValue("s", TO_WRAP.c_str()), chpl::UniqueString::get(&CONTEXT->context, PyUnicode_AsUTF8(TO_UNWRAP)));
-DEFINE_RETURN_TYPE(std::string, "str", Py_BuildValue("s", TO_WRAP.c_str()), std::string(PyUnicode_AsUTF8(TO_UNWRAP)));
-DEFINE_RETURN_TYPE(const chpl::uast::AstNode*, "AstNode", wrapAstNode(CONTEXT, TO_WRAP), ((AstNodeObject*) TO_UNWRAP)->astNode);
-DEFINE_RETURN_TYPE(IterAdapterBase*, "Iterator[AstNode]", wrapIterAdapter(CONTEXT, TO_WRAP), ((AstIterObject*) TO_UNWRAP)->iterAdapter);
-
-template<typename T> struct PythonFnHelper{};
-template<typename R, typename ...Args>
-struct PythonFnHelper<R(Args...)> {
-  using ReturnTypeInfo = PythonReturnTypeInfo<R>;
-  using ArgTypeInfo = std::tuple<PythonReturnTypeInfo<Args>...>;
-};
-
-static int ContextObject_init(ContextObject* self, PyObject* args, PyObject* kwargs) {
-  chpl::Context::Configuration config;
-  config.chplHome = getenv("CHPL_HOME");
-  new (&self->context) chpl::Context(std::move(config));
-  chpl::parsing::setupModuleSearchPaths(&self->context, false, false, {}, {});
-  return 0;
-}
-
-static void ContextObject_dealloc(ContextObject* self) {
-  self->context.~Context();
-  Py_TYPE(self)->tp_free((PyObject *) self);
-}
-
-static PyObject* ContextObject_parse(ContextObject *self, PyObject* args) {
-  auto context = &self->context;
-  const char* fileName;
-  if (!PyArg_ParseTuple(args, "s", &fileName)) {
-    PyErr_BadArgument();
-    return nullptr;
-  }
-  auto fileNameUS = chpl::UniqueString::get(context, fileName);
-  auto parentPathUS = chpl::UniqueString();
-  auto& builderResult = chpl::parsing::parseFileToBuilderResult(context, fileNameUS, parentPathUS);
-
-  int listSize = builderResult.numTopLevelExpressions();
-  PyObject* topExprs = PyList_New(listSize);
-  for (auto i = 0; i < listSize; i++) {
-    PyObject* node = wrapAstNode(self, builderResult.topLevelExpression(i));
-    PyList_SetItem(topExprs, i, node);
-  }
-  return topExprs;
-}
-
-static PyObject* ContextObject_is_bundled_path(ContextObject *self, PyObject* args) {
-  auto context = &self->context;
-  const char* fileName;
-  if (!PyArg_ParseTuple(args, "s", &fileName)) {
-    PyErr_BadArgument();
-    return nullptr;
-  }
-  auto pathUS = chpl::UniqueString::get(context, fileName);
-
-  bool isInternalPath =
-    chpl::parsing::filePathIsInInternalModule(context, pathUS) ||
-    chpl::parsing::filePathIsInStandardModule(context, pathUS) ||
-    chpl::parsing::filePathIsInBundledModule(context, pathUS);
-
-  return PyBool_FromLong(isInternalPath);
-}
-
-static PyObject* ContextObject_advance_to_next_revision(ContextObject *self, PyObject* args) {
-  auto context = &self->context;
-  bool prepareToGc;
-  if (!PyArg_ParseTuple(args, "b", &prepareToGc)) {
-    PyErr_BadArgument();
-    return nullptr;
-  }
-
-  context->advanceToNextRevision(prepareToGc);
-  Py_RETURN_NONE;
-}
-
-template <typename Tuple, size_t ... Indices>
-static void printFunctionArgs(std::ostringstream& ss, std::index_sequence<Indices...>) {
-  int counter = 0;
-  auto printArg = [&](const char* arg) {
-    ss << ", arg" << counter++ << ": " << arg;
-  };
-
-  int dummy[] = { (printArg(std::tuple_element<Indices, Tuple>::type::TypeString), 0)...};
-  (void) dummy;
-}
-
-static const char* tagToUserFacingStringTable[chpl::uast::asttags::NUM_AST_TAGS] = {
-// define tag to string conversion
-#define AST_NODE(NAME) #NAME,
-#define AST_LEAF(NAME) #NAME,
-#define AST_BEGIN_SUBCLASSES(NAME) #NAME,
-#define AST_END_SUBCLASSES(NAME) #NAME,
-// Apply the above macros to uast-classes-list.h
-#include "chpl/uast/uast-classes-list.h"
-// clear the macros
-#undef AST_NODE
-#undef AST_LEAF
-#undef AST_BEGIN_SUBCLASSES
-#undef AST_END_SUBCLASSES
-#undef NAMESTR
-};
-
-static PyObject* ContextObject_get_pyi_file(ContextObject *self, PyObject* args) {
-  std::ostringstream ss;
-
-  ss << "from typing import *" << std::endl << std::endl;
-
-  ss << "class AstNode:" << std::endl;
-  ss << "    pass" << std::endl << std::endl;
-
-  using namespace chpl;
-  using namespace uast;
-  #define CLASS_BEGIN(NODE) \
-    ss << "class " << tagToUserFacingStringTable[asttags::NODE] << "(AstNode):" << std::endl;
-  #define METHOD(NODE, NAME, DOCSTR, TYPEFN, BODY) \
-    ss << "    def " << #NAME << "(self"; \
-    printFunctionArgs<PythonFnHelper<TYPEFN>::ArgTypeInfo>(ss, std::make_index_sequence<std::tuple_size<PythonFnHelper<TYPEFN>::ArgTypeInfo>::value>()); \
-    ss << ") -> " << PythonFnHelper<TYPEFN>::ReturnTypeInfo::TypeString << ":" << std::endl;\
-    ss << "        \"\"\"" << std::endl; \
-    ss << "        " << DOCSTR << std::endl; \
-    ss << "        \"\"\"" << std::endl; \
-    ss << "        ..." << std::endl << std::endl;
-  #define CLASS_END(NODE) \
-    ss << std::endl;
-  #include "method-tables.h"
-
-  return Py_BuildValue("s", ss.str().c_str());
-}
-
-static PyMethodDef ContextObject_methods[] = {
-  { "parse", (PyCFunction) ContextObject_parse, METH_VARARGS, "Parse a top-level AST node from the given file" },
-  { "is_bundled_path", (PyCFunction) ContextObject_is_bundled_path, METH_VARARGS, "Check if the given file path is within the bundled (built-in) Chapel files" },
-  { "advance_to_next_revision", (PyCFunction) ContextObject_advance_to_next_revision, METH_VARARGS, "Advance the context to the next revision" },
-  { "get_pyi_file", (PyCFunction) ContextObject_get_pyi_file, METH_NOARGS, "Generate a stub file for the Chapel AST nodes" },
-  {NULL, NULL, 0, NULL}  /* Sentinel */
-};
-
-PyTypeObject ContextType = {
-  PyVarObject_HEAD_INIT(NULL, 0)
-  .tp_name = "Context",
-  .tp_basicsize = sizeof(ContextObject),
-  .tp_itemsize = 0,
-  .tp_dealloc = (destructor) ContextObject_dealloc,
-  .tp_flags = Py_TPFLAGS_DEFAULT,
-  .tp_doc = PyDoc_STR("The Chapel context object that tracks various frontend state"),
-  .tp_methods = ContextObject_methods,
-  .tp_init = (initproc) ContextObject_init,
-  .tp_new = PyType_GenericNew,
-};
-
-static int AstNodeObject_init(AstNodeObject* self, PyObject* args, PyObject* kwargs) { \
-  PyObject* contextObjectPy;
-  if (!PyArg_ParseTuple(args, "O", &contextObjectPy))
-      return -1;
-
-  Py_INCREF(contextObjectPy);
-  self->astNode = nullptr;
-  self->contextObject = contextObjectPy;
-  return 0;
-}
-
-static void AstNodeObject_dealloc(AstNodeObject* self) {
-  Py_XDECREF(self->contextObject);
-  Py_TYPE(self)->tp_free((PyObject *) self);
-}
-
-static PyObject* AstNodeObject_dump(AstNodeObject *self, PyObject *Py_UNUSED(ignored)) {
-  self->astNode->dump();
-  Py_RETURN_NONE;
-}
-
-static PyObject* AstNodeObject_tag(AstNodeObject *self, PyObject *Py_UNUSED(ignored)) {
-  const char* nodeType = chpl::uast::asttags::tagToString(self->astNode->tag());
-  return Py_BuildValue("s", nodeType);
-}
-
-static PyObject* AstNodeObject_unique_id(AstNodeObject *self, PyObject *Py_UNUSED(ignored)) {
-  auto uniqueID = (intptr_t)self->astNode;
-  return Py_BuildValue("K", uniqueID);
-}
-
-
-static PyObject* AstNodeObject_attribute_group(AstNodeObject *self, PyObject *Py_UNUSED(ignored)) {
-  return wrapAstNode((ContextObject*) self->contextObject,
-                     self->astNode->attributeGroup());
-}
-
-static PyObject* AstNodeObject_pragmas(AstNodeObject *self, PyObject *Py_UNUSED(ignored)) {
-  PyObject* elms = PySet_New(NULL);
-  auto attrs = self->astNode->attributeGroup();
-  if (attrs) {
-    for (auto p: attrs->pragmas()) {
-      PyObject* s = Py_BuildValue("s", chpl::uast::pragmatags::pragmaTagToName(p));
-      PySet_Add(elms, s);
-    }
-  }
-  return elms;
-}
-
-static PyObject* AstNodeObject_parent(AstNodeObject* self, PyObject *Py_UNUSED(ignored)) {
-  auto contextObject = (ContextObject*) self->contextObject;
-  auto context = &contextObject->context;
-
-  return wrapAstNode(contextObject, chpl::parsing::parentAst(context, self->astNode));
-}
-
-static PyObject* AstNodeObject_iter(AstNodeObject *self) {
-  return wrapIterPair((ContextObject*) self->contextObject, self->astNode->children());
-}
-
-static PyObject* AstNodeObject_location(AstNodeObject *self) {
-  auto locationObjectPy = PyObject_CallObject((PyObject *) &LocationType, nullptr);
-  auto& location = ((LocationObject*) locationObjectPy)->location;
-  auto context = &((ContextObject*) self->contextObject)->context;
-
-  location = chpl::parsing::locateAst(context, self->astNode);
-  return locationObjectPy;
-}
-
-static PyMethodDef AstNodeObject_methods[] = {
-  {"dump", (PyCFunction) AstNodeObject_dump, METH_NOARGS, "Dump the internal representation of the given AST node"},
-  {"tag", (PyCFunction) AstNodeObject_tag, METH_NOARGS, "Get a string representation of the AST node's type"},
-  {"attribute_group", (PyCFunction) AstNodeObject_attribute_group, METH_NOARGS, "Get the attribute group, if any, associated with this node"},
-  {"location", (PyCFunction) AstNodeObject_location, METH_NOARGS, "Get the location of this AST node in its file"},
-  {"parent", (PyCFunction) AstNodeObject_parent, METH_NOARGS, "Get the parent node of this AST node"},
-  {"pragmas", (PyCFunction) AstNodeObject_pragmas, METH_NOARGS, "Get the pragmas of this AST node"},
-  {"unique_id", (PyCFunction) AstNodeObject_unique_id, METH_NOARGS, "Get a unique identifer for this AST node"},
-  {NULL, NULL, 0, NULL} /* Sentinel */
-};
-
-PyTypeObject AstNodeType = {
-  PyVarObject_HEAD_INIT(NULL, 0)
-  .tp_name = "AstNode",
-  .tp_basicsize = sizeof(AstNodeObject),
-  .tp_itemsize = 0,
-  .tp_dealloc = (destructor) AstNodeObject_dealloc,
-  .tp_flags = Py_TPFLAGS_BASETYPE,
-  .tp_doc = PyDoc_STR("The base type of Chapel AST nodes"),
-  .tp_iter = (getiterfunc) AstNodeObject_iter,
-  .tp_methods = AstNodeObject_methods,
-  .tp_init = (initproc) AstNodeObject_init,
-  .tp_new = PyType_GenericNew,
-};
-
-static const char* blockStyleToString(chpl::uast::BlockStyle blockStyle) {
-  switch (blockStyle) {
-    case chpl::uast::BlockStyle::EXPLICIT: return "explicit";
-    case chpl::uast::BlockStyle::IMPLICIT: return "implicit";
-    case chpl::uast::BlockStyle::UNNECESSARY_KEYWORD_AND_BLOCK: return "unnecessary";
-  }
-}
-
-static const char* opKindToString(chpl::uast::Range::OpKind kind) {
-  switch (kind) {
-    case chpl::uast::Range::DEFAULT: return "..";
-    case chpl::uast::Range::OPEN_HIGH: return "..<";
-    default: return "";
-  }
-}
-
-
-template<typename IntentType>
-static const char* intentToString(IntentType intent) {
-  return chpl::uast::qualifierToString(chpl::uast::Qualifier(int(intent)));
-}
-
-template <typename T>
-const char* toCString(T& t) { throw std::invalid_argument("Invalid conversion into C string"); }
-
-template <>
-const char* toCString<const char*>(const char*& t) { return t; }
-
-template <>
-const char* toCString<chpl::UniqueString>(chpl::UniqueString& us) { return us.c_str(); }
-
-#define METHOD(NODE, NAME, DOCSTR, TYPEFN, BODY)\
-  static PyObject* NODE##Object_##NAME(PyObject *self, PyObject *Py_UNUSED(ignored)) {\
-    using namespace chpl; \
-    using namespace uast; \
-    \
-    auto cast = ((NODE##Object*) self)->parent.astNode->to##NODE(); \
-    auto contextObject = (ContextObject*) ((NODE##Object*) self)->parent.contextObject;\
-    auto result = [](const NODE* node) { \
-      BODY; \
-    }(cast) ; \
-    return PythonFnHelper<TYPEFN>::ReturnTypeInfo::wrap(contextObject, std::move(result));\
-  }
-#include "method-tables.h"
-
-#define ACTUAL_ITERATOR(NAME)\
-  static PyObject* NAME##Object_actuals(PyObject *self, PyObject *Py_UNUSED(ignored)) { \
-    auto node = ((NAME##Object*) self)->parent.astNode->to##NAME(); \
-    \
-    auto argList = Py_BuildValue("(O)", (PyObject*) self); \
-    auto astCallIterObjectPy = PyObject_CallObject((PyObject *) &AstCallIterType, argList); \
-    Py_XDECREF(argList); \
-    \
-    auto astCalliterObject = (AstCallIterObject*) astCallIterObjectPy; \
-    astCalliterObject->current = 0; \
-    astCalliterObject->num = node->numActuals(); \
-    astCalliterObject->container = node; \
-    astCalliterObject->childGetter = [](const void* n, int child) { \
-      return ((chpl::uast::NAME*) n)->actual(child); \
-    }; \
-    astCalliterObject->nameGetter = [](const void* n, int child) { \
-      return ((chpl::uast::NAME*) n)->actualName(child); \
-    }; \
-    \
-    return astCallIterObjectPy; \
-  }
-
-ACTUAL_ITERATOR(Attribute);
-ACTUAL_ITERATOR(FnCall);
-
-template <chpl::uast::asttags::AstTag tag>
-struct PerNodeInfo {
-  static constexpr PyMethodDef methods[] = {
-    {NULL, NULL, 0, NULL}  /* Sentinel */
-  };
-};
-
-#define CLASS_BEGIN(TAG) \
-  template <> \
-  struct PerNodeInfo<chpl::uast::asttags::TAG> { \
-    static constexpr PyMethodDef methods[] = {
-#define CLASS_END(TAG) \
-      {NULL, NULL, 0, NULL}  /* Sentinel */ \
-    }; \
-  };
-#define METHOD(NODE, NAME, DOCSTR, TYPE, BODY) \
-  {#NAME, NODE##Object_##NAME, METH_NOARGS, DOCSTR},
-#define METHOD_PROTOTYPE(NODE, NAME, DOCSTR) \
-  {#NAME, NODE##Object_##NAME, METH_NOARGS, DOCSTR},
-#include "method-tables.h"
-
-
-#define DEFINE_PY_TYPE_FOR(NAME, TAG, FLAGS)\
-  PyTypeObject NAME##Type = { \
-    PyVarObject_HEAD_INIT(NULL, 0) \
-    .tp_name = #NAME, \
-    .tp_basicsize = sizeof(NAME##Object), \
-    .tp_itemsize = 0, \
-    .tp_flags = FLAGS, \
-    .tp_doc = PyDoc_STR("A Chapel " #NAME " AST node"), \
-    .tp_methods = (PyMethodDef*) PerNodeInfo<TAG>::methods, \
-    .tp_base = parentTypeFor(TAG), \
-    .tp_init = (initproc) NAME##Object_init, \
-    .tp_new = PyType_GenericNew, \
-  }; \
-
-#define AST_NODE(NAME) DEFINE_PY_TYPE_FOR(NAME, chpl::uast::asttags::NAME, Py_TPFLAGS_DEFAULT)
-#define AST_LEAF(NAME) DEFINE_PY_TYPE_FOR(NAME, chpl::uast::asttags::NAME, Py_TPFLAGS_DEFAULT)
-#define AST_BEGIN_SUBCLASSES(NAME) DEFINE_PY_TYPE_FOR(NAME, chpl::uast::asttags::START_##NAME, Py_TPFLAGS_BASETYPE)
-#define AST_END_SUBCLASSES(NAME)
-#include "chpl/uast/uast-classes-list.h"
-#undef AST_NODE
-#undef AST_LEAF
-#undef AST_BEGIN_SUBCLASSES
-#undef AST_END_SUBCLASSES
 
 static PyMethodDef ChapelMethods[] = {
   { NULL, NULL, 0, NULL } /* Sentinel */
